Lots of tweeks and fixes with a test project

+++ TwitterApiExample/Repositories/IRepository.cs
@@ -4,6 +4,6 @@ public interface IRepository<T>
 {
     Task Save(T item);
     Task<IList<T>> GetAll();
-    Task<int> GetCount();
+    Task<long> GetCount();
 }
